backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Django",
    "djangorestframework",
    "django-cors-headers",
    "djangorestframework-simplejwt",
    "psycopg2-binary",
    "python-dotenv",
    "whitenoise",
    "redis",
    "django-redis",
    "kazoo",
    "markdown",
    "pygments",
]
